Made a different UI for the Revive Part and also Restructured the Canvas Heirarchies.
Changed UI for the Revive sequence, and now it uses DOTween for the animations instead of an IEnumerator function. Also all UI Canvases are now under one Root Canvas and its rendered in Camera Space instead of Overlay because custom materials for pulses at the revive screen werent working in Overlay.